fix: token count, cost calculation, and budget enforcement
- Inject stream_options.include_usage for OpenAI-compatible streaming
  so token counts and costs are tracked for streaming requests
- Fix f64→f32 precision loss: keep f64 throughout cost pipeline,
  cast to f32 only at FFI boundary for Swift UI
- Add contains-based fallback in pricing lookup: longest CSV key
  that is a substring of model name wins (e.g. gpt-4o matches
  gpt-4o-2024-11-20)
- Add server-side budget enforcement: return HTTP 429 when
  cost_controls.enabled and total cost >= budget (budget=0 blocks all)
- Change CostControlsConfig.budget from f32 to f64
- Add 429 Too Many Requests to HTTP status line mapping
- Skip upstream assertion in test runner when expected file absent
- Add case-10: budget exceeded integration test

10/10 tests passed.

+++ src/metrics.zig
@@ -74,13 +74,13 @@ pub fn addOutputTokens(tokens: u64) void {
 }
 
 /// Add input cost (in dollars, converted to micro-dollars internally)
-pub fn addInputCost(dollars: f32) void {
+pub fn addInputCost(dollars: f64) void {
     const micros: u64 = @intFromFloat(dollars * 1_000_000.0);
     _ = input_cost_micros.fetchAdd(micros, .monotonic);
 }
 
 /// Add output cost (in dollars, converted to micro-dollars internally)
-pub fn addOutputCost(dollars: f32) void {
+pub fn addOutputCost(dollars: f64) void {
     const micros: u64 = @intFromFloat(dollars * 1_000_000.0);
     _ = output_cost_micros.fetchAdd(micros, .monotonic);
 }
@@ -241,8 +241,8 @@ pub const Snapshot = struct {
     // LLM metrics
     input_tokens: u64,
     output_tokens: u64,
-    input_cost: f32,
-    output_cost: f32,
+    input_cost: f64,
+    output_cost: f64,
 };
 
 /// Get a consistent snapshot of all metrics including process stats
@@ -255,7 +255,7 @@ pub fn snapshot() Snapshot {
         .network_tx_bytes = network_tx_bytes.load(.monotonic),
         .input_tokens = input_tokens.load(.monotonic),
         .output_tokens = output_tokens.load(.monotonic),
-        .input_cost = @as(f32, @floatFromInt(input_cost_micros.load(.monotonic))) / 1_000_000.0,
-        .output_cost = @as(f32, @floatFromInt(output_cost_micros.load(.monotonic))) / 1_000_000.0,
+        .input_cost = @as(f64, @floatFromInt(input_cost_micros.load(.monotonic))) / 1_000_000.0,
+        .output_cost = @as(f64, @floatFromInt(output_cost_micros.load(.monotonic))) / 1_000_000.0,
     };
 }
